fix: avoid showing thumbnail in source
The old lazy-loading solution requires an extra data-src attribute where the real image url is set. The script will replace src with data-src value when img tag is observed.

But this caused a problem: Third-party applications like RSS readers do not load script, hence the image source is always the placeholder (a blurry version of the real image).

Considering the fact that a placeholder image does not improve user experience that well, I removed it.